+++ lib/commercetools-api/src/Client/Resource/ByProjectKeyProductProjectionsSearchGet.php
@@ -296,6 +296,15 @@ class ByProjectKeyProductProjectionsSearchGet extends ApiRequest implements Sort
 
     /**
      *
+     * @psalm-param scalar|scalar[] $priceCustomerGroupAssignments
+     */
+    public function withPriceCustomerGroupAssignments($priceCustomerGroupAssignments): ByProjectKeyProductProjectionsSearchGet
+    {
+        return $this->withQueryParam('priceCustomerGroupAssignments', $priceCustomerGroupAssignments);
+    }
+
+    /**
+     *
      * @psalm-param scalar|scalar[] $priceChannel
      */
     public function withPriceChannel($priceChannel): ByProjectKeyProductProjectionsSearchGet