+++ lib/commercetools-api/src/Models/Customer/CustomerRemoveCustomerGroupAssignmentAction.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Api\Models\Customer;
+
+use Commercetools\Api\Models\CustomerGroup\CustomerGroupResourceIdentifier;
+use Commercetools\Base\DateTimeImmutableCollection;
+use Commercetools\Base\JsonObject;
+
+interface CustomerRemoveCustomerGroupAssignmentAction extends CustomerUpdateAction
+{
+    public const FIELD_CUSTOMER_GROUP = 'customerGroup';
+
+    /**
+     * <p>Customer Group to unassign the Customer from.</p>
+     *
+
+     * @return null|CustomerGroupResourceIdentifier
+     */
+    public function getCustomerGroup();
+
+    /**
+     * @param ?CustomerGroupResourceIdentifier $customerGroup
+     */
+    public function setCustomerGroup(?CustomerGroupResourceIdentifier $customerGroup): void;
+}