+++ lib/commercetools-api/src/Models/Customer/CustomerRemoveCustomerGroupAssignmentActionModel.php
@@ -1,0 +1,96 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Api\Models\Customer;
+
+use Commercetools\Api\Models\CustomerGroup\CustomerGroupResourceIdentifier;
+use Commercetools\Api\Models\CustomerGroup\CustomerGroupResourceIdentifierModel;
+use Commercetools\Base\DateTimeImmutableCollection;
+use Commercetools\Base\JsonObject;
+use Commercetools\Base\JsonObjectModel;
+use Commercetools\Base\MapperFactory;
+use stdClass;
+
+/**
+ * @internal
+ */
+final class CustomerRemoveCustomerGroupAssignmentActionModel extends JsonObjectModel implements CustomerRemoveCustomerGroupAssignmentAction
+{
+    public const DISCRIMINATOR_VALUE = 'removeCustomerGroupAssignment';
+    /**
+     *
+     * @var ?string
+     */
+    protected $action;
+
+    /**
+     *
+     * @var ?CustomerGroupResourceIdentifier
+     */
+    protected $customerGroup;
+
+
+    /**
+     * @psalm-suppress MissingParamType
+     */
+    public function __construct(
+        ?CustomerGroupResourceIdentifier $customerGroup = null,
+        ?string $action = null
+    ) {
+        $this->customerGroup = $customerGroup;
+        $this->action = $action ?? self::DISCRIMINATOR_VALUE;
+    }
+
+    /**
+     *
+     * @return null|string
+     */
+    public function getAction()
+    {
+        if (is_null($this->action)) {
+            /** @psalm-var ?string $data */
+            $data = $this->raw(self::FIELD_ACTION);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->action = (string) $data;
+        }
+
+        return $this->action;
+    }
+
+    /**
+     * <p>Customer Group to unassign the Customer from.</p>
+     *
+     *
+     * @return null|CustomerGroupResourceIdentifier
+     */
+    public function getCustomerGroup()
+    {
+        if (is_null($this->customerGroup)) {
+            /** @psalm-var stdClass|array<string, mixed>|null $data */
+            $data = $this->raw(self::FIELD_CUSTOMER_GROUP);
+            if (is_null($data)) {
+                return null;
+            }
+
+            $this->customerGroup = CustomerGroupResourceIdentifierModel::of($data);
+        }
+
+        return $this->customerGroup;
+    }
+
+
+    /**
+     * @param ?CustomerGroupResourceIdentifier $customerGroup
+     */
+    public function setCustomerGroup(?CustomerGroupResourceIdentifier $customerGroup): void
+    {
+        $this->customerGroup = $customerGroup;
+    }
+}